refactor: step1 피드백 반영 - getStation() -> Sections로 책임이동
