+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -75,7 +75,7 @@ public class LineService {
     }
 
     public List<StationResponse> getStationResponse(Line line) {
-        return line.getStations().stream()
+        return line.getSections().getStations().stream()
                 .map(StationResponse::of)
                 .collect(Collectors.toList());
     }
